Move entry destroy logic into collection entry (#2124)
This PR refactors where entry destroy logic is run by moving it into the collection entry class itself instead of having that logic in the parent collection. This allows the collection to only be concerned with running the appropriate methods and lifecycle hooks instead of also removing owned properties of the entry. This PR also ensures that `onDeregisterEntry` is run after `destroyEntry` instead of before.