Create the ability to add plugin presets for specific modules (#2125)
With the new plugin system, there are some common configurations that are almost always applied depending on the module. For example, a drawer using `propStore` will most likely have this configuration:

```js
{
    prop: "inlineState",
    value: ({ entry }) => entry.store,
    condition: ({ entry }) => ["opened", "closed", "indeterminate"].includes(entry.state),
    onChange: ({ entry }) => entry.applyState()
}
```

This will set a prop watch on `inlineState` and store the state in local storage if a valid state is set. This PR adds these common configurations as module presets and a establishes a system for defining presets based on the module using the plugin.