Create new focusTrap plugin (#2129)
This PR converts the existing implementation of focus trapping into a modular plugin. This allows the feature to be excluded and added as needed to any component using the new plugin system and also enables more flexibility with the condition that it is applied. As part of this update, there's also a refactored `FocusTrap` class that can be used on its own where the `focusTrap` plugin implements the module in the form of a collection plugin.