Add breakpoint modifier support to bem function (#2135)
This PR adds a new argument to the core `bem()` function to add support for breakpoint modifiers. This feature has been carried over from similar implementation in the `utiltiy()` function which will prepend breakpoint modifiers to a class with a colon. This feature also ensure the colon is applied and escaped automatically.

**Additional changes**
- Applies new breakpoint modifier function to flex and grid components.
- Ensure `bem` is used for all flex and grid classes, even soft utilities.
- Fixed bug where block prefixes were being applied in front of breakpoint modifiers.
- Renamed breakpoint map loops to use `$bp` instead of `$key` for clarity.