Convert the modal package to TypeScript (#2375)
This PR migrates the `@vrembem/modal` package to typescript and enhances the build process to emit declaration files and modern ES modules. As part of this migration, the modal API has been simplified to be more predictable with its expected params and returned values. The core `StackArray` module has also been typed as a generic to properly accept `ModalEntry` types.