Remove aria-hidden from updateGlobalState module (#2420)
## What changed?

I'm currently getting this warning whenever the `updateGlobalState` module is used:

```
Blocked aria-hidden on an element because its descendant retained focus. The focus must not be hidden from assistive technology users. Avoid using aria-hidden on a focused element or its ancestor. Consider using the inert attribute instead, which will also prevent focus. For more details, see the aria-hidden section of the WAI-ARIA specification at https://w3c.github.io/aria/#aria-hidden.
```

This PR fixes this error by removing the use of `aria-hidden` because inert is enough to handle the global state of the page when an area needs to be disabled/hidden from user interaction. Tests are also fixed to accommodate this change.